refactor(LottoMachine.java): package구조 변경합니다
