+++ src/main/java/lotto/LottoApplication.java
@@ -1,5 +1,7 @@
 package lotto;
 
+import lotto.controller.LottoMachine;
+
 public class LottoApplication {
     public static void main(String[] args) {
         LottoMachine lottoMachine = new LottoMachine();
