+++ src/main/java/lotto/model/LottoNumbers.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model;
 
 import java.util.List;
 import java.util.Objects;
