+++ src/main/java/lotto/model/LottoQuantityChecker.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model;
 
 public class LottoQuantityChecker {
     private static final String INVALID_PRICE = "1000원 단위의 금액을 입력하세요";
