+++ src/test/java/lotto/LottoGeneratorTest.java
@@ -1,5 +1,6 @@
 package lotto;
 
+import lotto.model.LottoGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
