+++ src/test/java/lotto/LottoQuantityCheckerTest.java
@@ -1,5 +1,6 @@
 package lotto;
 
+import lotto.model.LottoQuantityChecker;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
