+++ src/test/java/lotto/LottoStaticsTest.java
@@ -1,13 +1,16 @@
 package lotto;
 
+import lotto.model.LottoNumbers;
+import lotto.model.LottoPrizeRanks;
+import lotto.model.LottoStatics;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
 import java.util.List;
 
-import static lotto.LottoPrizeRank.THREE;
-import static lotto.LottoPrizeRank.NONE;
+import static lotto.model.LottoPrizeRank.THREE;
+import static lotto.model.LottoPrizeRank.NONE;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @DisplayName("LottoStatics는 ")
