+++ src/test/java/lotto/RandomNumberGeneratorTest.java
@@ -1,5 +1,6 @@
 package lotto;
 
+import lotto.model.RandomNumberGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
